Update package structure by renaming the strategy package to comparison and moving classes regarding the comparison algorithm from the main jplag package into the comparison package.
